Created GitHub Documentation (#1)
* Generated markdown documentation using the PlatyPS module. This is the initial revision that will need cleanup.

Signed-off-by: Paul Boyer <pboyer2@wisc.edu>

* Updated the summary page.

Signed-off-by: Paul Boyer <pboyer2@wisc.edu>

* Updated Location of Function Comments

Edited the module file to remove unecessary whitespace as well as
relocate the comment-based help inside of the funciton declaration. Also
modified the UpdateAPIFunctions.ps1 file to insert the comment-based
help inside of the function definition when generating the updated API
cmdlets from the API documentation.

Signed-off-by: Paul Boyer <pboyer2@wisc.edu>

* Updated PS-Doc Template and ReadMe

Updated the template that PS-doc uses to create documentation (for
markdown). This was used to then generate the new ReadMe file. The
contents of the old ReadMe have been moved to "Module Summary.md". The
older documentation files have been removed.

Signed-off-by: Paul Boyer <pboyer2@wisc.edu>

* Finished Updating Metadata

Finished updating metadata such as the module version and the copyright
year.

Signed-off-by: Paul Boyer <pboyer2@wisc.edu>