Update to O365 API Function Import
Added a clause to the Update/UpdateAPIFunctions.ps1 file that will
instruct users that this function needs to be elevated when executing to
work properly (at least the first time when trying to update/generate
documentation). Also added a suppression of an invalid module error
thrown when this function is invoked from the WiscO365.psm1 file.

Signed-off-by: Paul Boyer <pboyer2@wisc.edu>